Add forum read status API and is_viewed property
Introduces a new API endpoint to mark a forum as read for a user and updates forum fetching utilities to include an is_viewed property on the last_chat object. Also adds an example file demonstrating the forum read status feature and how the is_viewed property can be used in the frontend to indicate unread messages.